+++ _templates/baseComponent/new/stories.ejs.t
@@ -2,7 +2,7 @@
 to: src/components/base/<%=name%>/<%=name%>.stories.tsx
 ---
 import React from "react";
-import { Meta, StoryObj } from "@storybook/react";
+import type { Meta, StoryObj } from "@storybook/react";
 import <%=name%> from "./<%=name%>";
 import StorybookContainer from "../StorybookContainer/StorybookContainer";
 
@@ -15,3 +15,12 @@ const meta: Meta<typeof <%=name%>> = {
 export default meta;
 
 type Story = StoryObj<typeof <%=name%>>;
+
+export const Default: Story = {
+  args: {},
+  render: (args) => (
+    <StorybookContainer>
+      <<%=name%> {...args} />
+    </StorybookContainer>
+  ),
+};